fix: create router files and add AccountLayout
