+++ src/pages/AccountPage.tsx
@@ -1,64 +1,107 @@
-import { NavLink, Outlet } from 'react-router-dom';
-import Footer from '@/components/our-components/footer';
-import Header from '@/components/our-components/header';
-import ActionButton from '@/components/our-components/actionButton';
+import { useState } from 'react';
+import { Pencil, Check } from 'lucide-react';
 
-interface SidebarItem {
+interface User {
   name: string;
-  path: string;
+  phone: string;
+  email: string;
+  description: string;
+  skills: string;
 }
 
-const SidebarMenu: SidebarItem[] = [
-  { name: 'Profile', path: '/profile' },
-  { name: 'Account Setting', path: '/account-setting' },
-  { name: 'Privacy', path: '/privacy' },
-  { name: 'Your Request', path: '/request' },
-  { name: 'Your Service', path: '/service' },
-];
-
 export default function AccountPage() {
+  const [user, setUser] = useState<User>({
+    name: 'John Doe',
+    phone: '+00 000 000 000',
+    email: 'john@doe.com',
+    description: 'I do some cleaning',
+    skills: 'I did some cleaning',
+  });
+
+  const [editingField, setEditingField] = useState<keyof User | null>(null);
+  const [tempValue, setTempValue] = useState<string>('');
+
+  const startEditing = (field: keyof User, value: string) => {
+    setEditingField(field);
+    setTempValue(value);
+  };
+
+  const saveEditing = () => {
+    if (editingField) {
+      setUser({ ...user, [editingField]: tempValue });
+      setEditingField(null);
+    }
+  };
+
+  const cancelEditing = () => {
+    setEditingField(null);
+    setTempValue('');
+  };
+
+  // Props type for the inner component
+  interface EditableFieldProps {
+    label: string;
+    field: keyof User;
+  }
+
+  const EditableField = ({ label, field }: EditableFieldProps) => (
+    <div>
+      <span className="font-semibold flex items-center">
+        {label}
+        {editingField !== field && (
+          <Pencil
+            onClick={() => startEditing(field, user[field])}
+            className="w-4 h-4 ml-2 text-gray-500 cursor-pointer hover:text-gray-700"
+          />
+        )}
+      </span>
+
+      {editingField === field ? (
+        <div className="flex items-center gap-2 mt-1">
+          <input
+            type="text"
+            value={tempValue}
+            onChange={(e) => setTempValue(e.target.value)}
+            className="border rounded px-2 py-1 flex-1"
+            autoFocus
+            onBlur={cancelEditing}
+            onKeyDown={(e) => {
+              if (e.key === 'Enter') saveEditing();
+              if (e.key === 'Escape') cancelEditing();
+            }}
+          />
+          <Check
+            onClick={saveEditing}
+            className="w-4 h-4 text-green-600 cursor-pointer"
+          />
+        </div>
+      ) : (
+        <p>{user[field]}</p>
+      )}
+    </div>
+  );
+
   return (
     <>
-      <Header />
-      <div className="w-full min-h-screen h-fit px-12 py-8 flex gap-6 bg-gray-50 justify-center">
-        {/* Sidebar */}
-        <div className="flex flex-col w-1/4 min-w-[160px] max-w-[240px]">
-          <h1 className="text-2xl font-bold mb-2">Your Account</h1>
-          <div className="w-full h-full bg-white border rounded-2xl px-4 pb-4 pt-8 flex flex-col gap-4 shadow-sm text-center">
-            {SidebarMenu.map((item: SidebarItem) => (
-              <>
-                <NavLink
-                  key={item.path}
-                  to={'/account' + item.path}
-                  className="font-medium hover:text-button-action"
-                >
-                  {item.name}
-                </NavLink>
-                {(item.name === 'Privacy' || item.name === 'Your Service') && (
-                  <hr className="my-2" />
-                )}
-              </>
-            ))}
-            <div className="mt-auto">
-              <ActionButton
-                buttonColor="red"
-                buttonType="outline"
-                onClick={() => alert('Logging out...')}
-              >
-                Log out
-              </ActionButton>
-            </div>
-          </div>
+      {/* Header */}
+      <h1 className="text-2xl font-bold mb-2">Profile</h1>
+
+      {/* Content */}
+      <div className="w-full h-full flex flex-col items-center bg-white border rounded-2xl p-8 shadow-sm m-0">
+        {/* Avatar */}
+        <div className="w-20 h-20 rounded-full bg-gray-100 flex items-center justify-center mb-6">
+          <Pencil className="w-6 h-6 text-gray-500" />
         </div>
-        {/* Main Content */}
-        <div className="w-full flex flex-col flex-1 max-w-[900px]">
-          <h1 className="text-2xl font-bold mb-2">Profile</h1>
-          <div className="w-full h-full bg-white border rounded-2xl py-8 px-4 shadow-sm m-0">
-            <Outlet />
-          </div>
+
+        {/* Editable Fields */}
+        <div className="space-y-4 w-full">
+          <EditableField label="Name" field="name" />
+          <EditableField label="Telephone" field="phone" />
+          <EditableField label="Email" field="email" />
+          <EditableField label="Description" field="description" />
+          <EditableField label="Skill & Experience" field="skills" />
         </div>
       </div>
-      <Footer />
     </>
   );
 }
