+++ src/pages/YourServicePage.tsx
@@ -1,0 +1,13 @@
+export default function YourServicePage() {
+  return (
+    <>
+      {/* Header */}
+      <h1 className="text-2xl font-bold mb-2">Your Service</h1>
+
+      {/* Content */}
+      <div className="w-full h-full flex flex-col items-center bg-white border rounded-2xl p-8 shadow-sm m-0">
+        <p>This is Your Service Page</p>
+      </div>
+    </>
+  );
+}
